confirmation_instructions: add "do not reply" remark to the email body
There have been several replies to this email containing nothing but the
sentence `Yes, I confirm.` or just simply `Confirmed`.  It seems like
someone discovered that replying to the email with exactly that helped
with recovering their Twitter account after it was locked, and now people
try to do this everywhere else.